move `startRemoteProxySession` into its own module so it is possible to mock it
