+++ packages/wrangler/src/api/remoteBindings/index.ts
@@ -1,94 +1,19 @@
 import assert from "node:assert";
-import path from "node:path";
-import getPort from "get-port";
-import remoteBindingsWorkerPath from "worker:remoteBindings/ProxyServerWorker";
 import { readConfig } from "../../config";
 import { getCloudflareComplianceRegion } from "../../environment-variables/misc-variables";
-import { getBasePath } from "../../paths";
 import { requireApiToken, requireAuth } from "../../user";
-import {
-	convertConfigBindingsToStartWorkerBindings,
-	startWorker,
-} from "../startDevWorker";
+import { convertConfigBindingsToStartWorkerBindings } from "../startDevWorker";
+import { startRemoteProxySession } from "./start-remote-proxy-session";
 import type { Config } from "../../config";
 import type { CfAccount } from "../../dev/create-worker-preview";
 import type {
 	AsyncHook,
 	Binding,
 	StartDevWorkerInput,
-	Worker,
 } from "../startDevWorker/types";
-import type { RemoteProxyConnectionString } from "miniflare";
+import type { RemoteProxySession } from "./start-remote-proxy-session";
 
-export type RemoteProxySession = Pick<Worker, "ready" | "dispose"> & {
-	updateBindings: (bindings: StartDevWorkerInput["bindings"]) => Promise<void>;
-	remoteProxyConnectionString: RemoteProxyConnectionString;
-};
-
-export type StartRemoteProxySessionOptions = {
-	workerName?: string;
-	auth?: NonNullable<StartDevWorkerInput["dev"]>["auth"];
-	/** If running in a non-public compliance region, set this here. */
-	complianceRegion?: Config["compliance_region"];
-};
-
-export async function startRemoteProxySession(
-	bindings: StartDevWorkerInput["bindings"],
-	options?: StartRemoteProxySessionOptions
-): Promise<RemoteProxySession> {
-	// Transform all bindings to use "raw" mode
-	const rawBindings = Object.fromEntries(
-		Object.entries(bindings ?? {}).map(([key, binding]) => [
-			key,
-			{ ...binding, raw: true },
-		])
-	);
-
-	const proxyServerWorkerWranglerConfig = path.resolve(
-		getBasePath(),
-		"templates/remoteBindings/wrangler.jsonc"
-	);
-
-	const worker = await startWorker({
-		name: options?.workerName,
-		entrypoint: remoteBindingsWorkerPath,
-		config: proxyServerWorkerWranglerConfig,
-		compatibilityDate: "2025-04-28",
-		dev: {
-			remote: "minimal",
-			auth: options?.auth,
-			server: {
-				port: await getPort(),
-			},
-			inspector: false,
-			logLevel: "error",
-		},
-		bindings: rawBindings,
-	});
-
-	const remoteProxyConnectionString =
-		(await worker.url) as RemoteProxyConnectionString;
-
-	const updateBindings = async (
-		newBindings: StartDevWorkerInput["bindings"]
-	) => {
-		// Transform all new bindings to use "raw" mode
-		const rawNewBindings = Object.fromEntries(
-			Object.entries(newBindings ?? {}).map(([key, binding]) => [
-				key,
-				{ ...binding, raw: true },
-			])
-		);
-		await worker.patchConfig({ bindings: rawNewBindings });
-	};
-
-	return {
-		ready: worker.ready,
-		remoteProxyConnectionString,
-		updateBindings,
-		dispose: worker.dispose,
-	};
-}
+export * from "./start-remote-proxy-session";
 
 export function pickRemoteBindings(
 	bindings: Record<string, Binding>
